No cache (#964)
### Changes
- Enable not using a verification result cache, using the new class
`EmptyVerificationResultCache`

### Testing
- Manually tried out using EmptyVerificationResultCache in Dafny, which
reduces memory pressure
